op-cancel-appointment - move bark-report into its own model file
